requirements.txt
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
distro==1.9.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.4
fuzzywuzzy==0.18.0
h11==0.14.0
htmltools==0.5.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jieba3k==0.35.1
Jinja2==3.1.3
joblib==1.4.0
Levenshtein==0.25.1
linkify-it-py==2.0.3
lxml==5.2.1
lxml_html_clean==0.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdit-py-plugins==0.4.0
mdurl==0.1.2
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
openai==1.19.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
prompt-toolkit==3.0.36
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-Levenshtein==0.25.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
questionary==2.0.1
rapidfuzz==3.8.1
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
rsconnect_python==1.23.0
semver==2.13.0
sgmllib3k==1.0.0
shiny==0.8.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
tinysegmenter==0.3
tldextract==5.1.2
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
